+++ sandy/__init__.py
@@ -31,6 +31,7 @@ import sandy.sampling
 from .samples import *
 from .records import *
 from .lpc import *
+from .cov import *
 
 testdir = "tests"
