+++ sandy/core/__init__.py
@@ -1,3 +1,2 @@
 from .endf6 import *
 from .xs import *
-from .cov import *
