+++ sandy/errorr.py
@@ -3,7 +3,7 @@ import numpy as np
 import logging
 
 from .core.endf6 import _FormattedFile
-from .core.cov import CategoryCov
+from .cov import CategoryCov
 from .core.xs import Xs
 from .records import read_cont, read_list
